Customization of CaptureCards page

+++ AmeCommon/CardsCapture/AddResourcesToProjectCommand.cs
@@ -9,16 +9,18 @@ namespace AmeCommon.CardsCapture
     {
         public string Mp3FolderName { get; }
         private readonly IHostingEnvironment environment;
+        private readonly AmeFotoVideoProject project;
         private readonly DirectoryInfo destinationDirectory;
         public override string Name => "Tworzenie plików dodatkowych projektu";
         public override string Label => $"Template - {Mp3FolderName}";
 
 
-        public AddResourcesToProjectCommand(string mp3FolderName, IHostingEnvironment environment, DirectoryInfo destinationDirectory, IDriveManager driveManager) : base(driveManager)
+        public AddResourcesToProjectCommand(string mp3FolderName, IHostingEnvironment environment, AmeFotoVideoProject project, IDriveManager driveManager) : base(driveManager)
         {
             Mp3FolderName = mp3FolderName;
             this.environment = environment;
-            this.destinationDirectory = destinationDirectory;
+            this.project = project;
+            destinationDirectory = project.GetLocalPathRoot();
             Device = new Device {UniqueName = "Startup"};
             SourceDrive = new DriveInfo(Path.GetPathRoot(environment.WebRootPath) ?? "c:");
             Commands = new List<MoveFileCommand>();
@@ -28,19 +30,26 @@ namespace AmeCommon.CardsCapture
         {
             CopyFile("nadruk_plyta.psd");
             CopyFile("nadruk_pudelko.psd");
-            CopyFile("temp.prproj");
+            CopyFile("temp.prproj", $"{project.Name}.prproj");
 
-            var mp3FilesDirPath = Path.Combine(environment.WebRootPath, "Resources", Mp3FolderName);
             var podkladyDestPath = Path.Combine(destinationDirectory.FullName, "Podkłady");
             Directory.CreateDirectory(podkladyDestPath);
-            foreach(var podklad in Directory.GetFiles(mp3FilesDirPath))
-                File.Copy(podklad, Path.Combine(podkladyDestPath, Path.GetFileName(podklad) ?? ""));
+
+            if (!string.IsNullOrEmpty(Mp3FolderName))
+            {
+                var mp3FilesDirPath = Path.Combine(environment.WebRootPath, "Resources", Mp3FolderName);
+                if (Directory.Exists(mp3FilesDirPath))
+                {
+                    foreach (var podklad in Directory.GetFiles(mp3FilesDirPath))
+                        File.Copy(podklad, Path.Combine(podkladyDestPath, Path.GetFileName(podklad) ?? ""));
+                }
+            }
         }
 
-        private void CopyFile(string fileName)
+        private void CopyFile(string fileName, string destFileName = null)
         {
             var sourceFile = Path.Combine(environment.WebRootPath, "Resources", fileName);
-            var destFile = Path.Combine(destinationDirectory.FullName, fileName);
+            var destFile = Path.Combine(destinationDirectory.FullName, destFileName ?? fileName);
             File.Copy(sourceFile, destFile);
         }
